Remove confusing EUCALYPTUS_CERT references
It was required a long time ago. For details, see:
https://bugs.launchpad.net/ubuntu/+source/euca2ools/+bug/479836
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=709475

The default is now /usr/share/euca2ools/cert-ec2.pem which is deployed
as part of the euca2ools installation. The user may still export the
environment variable to override this default if required, but there
is no longer any point in checking the environment variable is
present.