Documentation updates
Mostly consists of formatting changes, but also drops references to
Wheezy AMIs and GCE, both of which are now unsupported. The
regression-checklist.md list has also been pruned of unnecessary items
(eg. system log, motd and eth0 checks which are not applicable since
at least Jessie).

I also added a warning to README.md that S3 buckets need to be located
in the same region as the instance doing the build.