Use msdos partitioning by default
On one hand, this is better supported by GRUB, which is applicable to
modern HVM instance types. On the other hand, this is incompatible
with paravirtual instances using instance-backed storage and will
require a template or argument to specify "none" infor those AMI
types.

Since most people will be building HVM AMIs these days, setting msdos
as the default makes sense.

Note: I have not tested cloud-init's ability to resize a root volume,
if partitioning is used, but it shouldn't be too difficult for
cloud-init to address if there's a problem. The user can always
manually set partitioning=none if needed.