Modify (not clobber) the cloud-init cloud.cfg
When upgrading cloud-init, some new options are expected (such as a
"security" repository URL under package_mirrors). If the settings are
missing, certain functionality will either break or print warnings.

Instead of tracking a separate configuration for the different
versions of cloud-init we might want to use, it is easier (and
probably safer) to just strip out the parts of the configuration which
we know we don't need.

Since cloud.cfg is written in YAML, it was very difficult to modify
the options accurately and reliably within Bash; hence the
introduction of a new dependency on Python. The new Python code should
work on any reasonably modern 2 or 3 release, and was tested against
2.7.9 and 3.4.2 specifically since those are found in Jessie.