Improve Grub handling
grub-xen doesn't actually work on EC2 paravirtual instances, since the
AKIs Amazon provides containing pv-grub are actually based on Grub
version 1 - an ancient and completely incompatible format!

Fortunately, Debian provides the pv-grub-menu package to be able to
satisfy the old pv-grub requirement. All it does is deploy
/boot/grub/menu.lst along with the update-menu-lst command. All
configuration is done directly within menu.lst comments, and
update-menu-lst still needs to be executed to update the format of
uncommented lines which follow (and are not to be edited directly) -
hence all the new regexes.

No matter the instance type, the /dev/sd? block device naming
convention is never used on modern Jessie EC2 instances, and in fact
will cause failures in a number of situations. The "sanity check"
block is not expected to ever be used, but could conceivably be
executed if debian-image-builder were called from a different and/or
sufficiently old host type.