Configure timezone using dpkg-reconfigure tzdata
We should end up with a more standard environment if we rely on the
Debian tooling when we can. However in this case, there are a few
catches.

Jessie seems to require /etc/timezone be set prior to running
dpkg-reconfigure, and the result will be /etc/localtime copied from
the applicable zoneinfo file. If a symlink already exists at
/etc/localtime, it will be replaced.

Stretch behaves similar to Ubuntu 16.04 (see the bug report link in
the diff) whereby the contents of /etc/timezone is ignored and
overwritten unconditionally, and /etc/localtime is expected to be a
symlink to the applicable zoneinfo file and used to detect the correct
timezone for writing to /etc/timezone.

To address these discrepancies while still permitting dpkg-reconfigure
to execute correctly on both Debian releases, we take care of creating
both /etc/timezone (with the correct timezone) and /etc/localtime (as
a symlink).