Simplify install-euca2ools task
We no longer attempt to check the euca2ools version. The test was
never very reliable since euca-version doesn't print the full semantic
version string.

We now assume that if we're running Debian stretch and have euca2ools
installed, it should be at least as new as the version Stretch
packages and we'll be good. If not running at least Stretch and don't
have euca2ools installed, only then will we proceed to build and
install it.

The only potential problem here is if the user is running a Jessie AMI
and has installed euca2ools via apt-get manually (or it was baked into
the AMI). This seems unlikely, but it'll be the fault of the user for
running without meeting the minimum documented requirements.

Also we now only create a temporary directory if required.