Configure grub.cfg systemd console/serial output
When plymouth is installed, this will enable systemd output to print
messages to both tty0 and ttyS0, which is useful since a console
screenshot in AWS is available much quicker immediately after launch
than fetching the serial output.

It's also useful if the host was rebooted since AWS won't update the
serial output if the event wasn't triggered through the AWS APIs. AWS
only captures serial output for a few minutes after an instance
starts.

The changes here also enable GRUB to print error messages over serial
shoud that ever be helpful, but in general this won't be obvious since
we have GRUB_TIMEOUT set to 0.